+++ ext/clouddrives-gdrive/webapp/src/main/webapp/WEB-INF/conf/configuration.xml
@@ -24,103 +24,4 @@
   xsi:schemaLocation="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd http://www.exoplatform.org/xml/ns/kernel_1_2.xsd"
   xmlns="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd">
 
-  <!-- Google Drive namespace for use within JCR -->
-  <!--<external-component-plugins>
-    <target-component>org.exoplatform.services.jcr.RepositoryService</target-component>
-    <component-plugin>
-      <name>add.namespaces</name>
-      <set-method>addPlugin</set-method>
-      <type>org.exoplatform.services.jcr.impl.AddNamespacesPlugin</type>
-      <init-params>
-        <properties-param>
-          <name>namespaces</name>
-          <property name="gdrive" value="http://www.exoplatform.com/jcr/googledrive/1.0/" />
-        </properties-param>
-      </init-params>
-    </component-plugin>
-  </external-component-plugins>-->
-
-  <!-- Google Drive connector plugin -->
-  <!--<external-component-plugins>
-    <target-component>org.exoplatform.services.cms.clouddrives.CloudDriveService</target-component>
-    <component-plugin>
-      <name>add.clouddriveprovider</name>
-      <set-method>addPlugin</set-method>
-      <type>org.exoplatform.services.cms.clouddrives.gdrive.GoogleDriveConnector</type>
-      <init-params>
-        <properties-param>
-          <name>drive-configuration</name>
-          <property name="provider-id" value="gdrive" />
-          <property name="provider-name" value="Google Drive" />
-          <property name="provider-client-id" value="${clouddrive.google.client.id}" />
-          <property name="provider-client-secret" value="${clouddrive.google.client.secret}" />
-          <property name="connector-host" value="${clouddrive.service.host}" />
-          <property name="connector-schema" value="${clouddrive.service.schema:http}" />
-          <property name="disable" value="${clouddrive.google.disable:false}" />
-        </properties-param>
-      </init-params>
-    </component-plugin>
-  </external-component-plugins>-->
-
-  <!--
-  Dedicated ECMS menu action for Google Drive.
-  Note that Google Drive will be added automatically to Connect Cloud Documents dialog
-  -->
-  <!-- PLF 6.1 Task 26787.
-  Hide connect google drive from explorer by commenting ShowConnectCloudDrive -->
-  <!--<external-component-plugins>
-    <target-component>org.exoplatform.webui.ext.UIExtensionManager</target-component>
-    <component-plugin>
-      <name>Add CloudDrive Actions</name>
-      <set-method>registerUIExtensionPlugin</set-method>
-      <type>org.exoplatform.webui.ext.UIExtensionPlugin</type>
-      <init-params>
-        <object-param>
-          <name>ConnectGDrive</name>
-          <object type="org.exoplatform.webui.ext.UIExtension">
-            <field name="type">
-              <string>org.exoplatform.ecm.dms.UIActionBar</string>
-            </field>
-            <field name="name">
-              <string>ConnectGDrive</string>
-            </field>
-            <field name="component">
-              <string>org.exoplatform.services.cms.clouddrives.gdrive.ecms.ConnectGDriveActionComponent</string>
-            </field>
-            <field name="extendedFilters">
-              <collection type="java.util.ArrayList">
-                <value>
-                  <object type="org.exoplatform.services.cms.clouddrives.webui.filters.PersonalDocumentsFilter"></object>
-                </value>
-              </collection>
-            </field>
-          </object>
-        </object-param>
-      </init-params>
-    </component-plugin>
-  </external-component-plugins>-->
-
-  <!--
-  ResourceBundleService with texts for Google Drive UI support.
-  -->
-  <!--<external-component-plugins>
-    <target-component>org.exoplatform.services.resources.ResourceBundleService</target-component>
-    <component-plugin>
-      <name>ResourceBundle Plugin</name>
-      <set-method>addResourceBundle</set-method>
-      <type>org.exoplatform.services.resources.impl.BaseResourceBundlePlugin</type>
-      <init-params>
-        <values-param>
-          <name>init.resources</name>
-          <description>Store the following resources into the db for the first launch</description>
-          <value>locale.clouddrive.CloudDrive-gdrive</value>
-        </values-param>
-        <values-param>
-          <name>portal.resource.names</name>
-          <description>The properties files of the portal , those file will be merged into one ResoruceBundle properties</description>
-          <value>locale.clouddrive.CloudDrive-gdrive</value>
-        </values-param>
-      </init-params>
-    </component-plugin>
-  </external-component-plugins>-->
 </configuration>
