+++ ext/clouddrives-gdrive/webapp/src/main/webapp/WEB-INF/gatein-resources.xml
@@ -24,17 +24,4 @@
     xsi:schemaLocation="http://www.gatein.org/xml/ns/gatein_resources_1_3 http://www.gatein.org/xml/ns/gatein_resources_1_3"
     xmlns="http://www.gatein.org/xml/ns/gatein_resources_1_3">
 
-  <!-- Javascript for Cloud Drive support of Google Drive -->
-  <!--<module>
-    <name>cloudDrive_gdrive</name>
-    <script>
-      <path>/js/clouddrive-gdrive.js</path>
-    </script>
-    <depends>
-      <module>jquery</module>
-    </depends>
-    <depends>
-      <module>cloudDriveUtils</module>
-    </depends>
-  </module>-->
 </gatein-resources>
