+++ ext/clouddrives-onedrive/webapp/src/main/webapp/WEB-INF/conf/configuration.xml
@@ -24,103 +24,4 @@
   xsi:schemaLocation="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd http://www.exoplatform.org/xml/ns/kernel_1_2.xsd"
   xmlns="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd">
 
-  <!-- One Drive namespace for use within JCR -->
-  <!--<external-component-plugins>
-    <target-component>org.exoplatform.services.jcr.RepositoryService</target-component>
-    <component-plugin>
-      <name>add.namespaces</name>
-      <set-method>addPlugin</set-method>
-      <type>org.exoplatform.services.jcr.impl.AddNamespacesPlugin</type>
-      <init-params>
-        <properties-param>
-          <name>namespaces</name>
-          <property name="onedrive" value="http://www.exoplatform.com/jcr/onedrive/1.0/" />
-        </properties-param>
-      </init-params>
-    </component-plugin>
-  </external-component-plugins>-->
-
-  <!-- One Drive connector plugin -->
-  <!--<external-component-plugins>
-    <target-component>org.exoplatform.services.cms.clouddrives.CloudDriveService</target-component>
-    <component-plugin>
-      <name>add.clouddriveprovider</name>
-      <set-method>addPlugin</set-method>
-      <type>org.exoplatform.services.cms.clouddrives.onedrive.OneDriveConnector</type>
-      <init-params>
-        <properties-param>
-          <name>drive-configuration</name>
-          <property name="provider-id" value="onedrive" />
-          <property name="provider-name" value="One Drive" />
-          <property name="provider-client-id" value="${clouddrive.onedrive.client.id}" />
-          <property name="provider-client-secret" value="${clouddrive.onedrive.client.secret}" />
-          <property name="connector-host" value="${clouddrive.service.host}" />
-          <property name="connector-schema" value="${clouddrive.service.schema:http}" />
-          <property name="disable" value="${clouddrive.onedrive.disable:false}" />
-        </properties-param>
-      </init-params>
-    </component-plugin>
-  </external-component-plugins>-->
-
-  <!--
-  Dedicated ECMS menu action for One Drive.
-  Note that One Drive will be added automatically to Connect Cloud Documents dialog
-  -->
-  <!-- PLF 6.1 Task 26787.
-  Hide connect OneDrive from explorer by commenting ShowConnectCloudDrive -->
-  <!--<external-component-plugins>
-    <target-component>org.exoplatform.webui.ext.UIExtensionManager</target-component>
-    <component-plugin>
-      <name>Add CloudDrive Actions</name>
-      <set-method>registerUIExtensionPlugin</set-method>
-      <type>org.exoplatform.webui.ext.UIExtensionPlugin</type>
-      <init-params>
-        <object-param>
-          <name>ConnectOneDrive</name>
-          <object type="org.exoplatform.webui.ext.UIExtension">
-            <field name="type">
-              <string>org.exoplatform.ecm.dms.UIActionBar</string>
-            </field>
-            <field name="name">
-              <string>ConnectOneDrive</string>
-            </field>
-            <field name="component">
-              <string>org.exoplatform.services.cms.clouddrives.onedrive.ecms.ConnectOneDriveActionComponent</string>
-            </field>
-            <field name="extendedFilters">
-              <collection type="java.util.ArrayList">
-                <value>
-                  <object type="org.exoplatform.services.cms.clouddrives.webui.filters.PersonalDocumentsFilter"></object>
-                </value>
-              </collection>
-            </field>
-          </object>
-        </object-param>
-      </init-params>
-    </component-plugin>
-  </external-component-plugins>-->
-
-  <!--
-  ResourceBundleService with texts for One Drive UI support.
-  -->
-  <!--<external-component-plugins>
-    <target-component>org.exoplatform.services.resources.ResourceBundleService</target-component>
-    <component-plugin>
-      <name>ResourceBundle Plugin</name>
-      <set-method>addResourceBundle</set-method>
-      <type>org.exoplatform.services.resources.impl.BaseResourceBundlePlugin</type>
-      <init-params>
-        <values-param>
-          <name>init.resources</name>
-          <description>Store the following resources into the db for the first launch</description>
-          <value>locale.clouddrive.CloudDrive-onedrive</value>
-        </values-param>
-        <values-param>
-          <name>portal.resource.names</name>
-          <description>The properties files of the portal , those file will be merged into one ResoruceBundle properties</description>
-          <value>locale.clouddrive.CloudDrive-onedrive</value>
-        </values-param>
-      </init-params>
-    </component-plugin>
-  </external-component-plugins>-->
 </configuration>
